+++ lib/pah/partials/_default.rb
@@ -6,20 +6,11 @@ copy_static_file 'app/views/application/_error_messages.html.haml'
 copy_static_file 'app/views/application/_flash_messages.html.haml'
 copy_static_file 'config/unicorn.rb'
 copy_static_file 'Procfile'
-copy_static_file "lib/tasks/integration.rake"
-
 copy_static_file 'config/locales/pt-BR.yml'
-
-gsub_file 'lib/tasks/integration.rake', /PROJECT/, @app_name
-
 copy_static_file '.gitignore'
 
-create_file ".env" do
-<<-EOF
-export PRODUCTION_APP=#{@app_name}
-export STAGING_APP=#{@app_name}-staging
-EOF
-end
+create_file ".env"
+append_to_file '.env', "APP: #{@app_name}\n"
 
 git :add => '.'
 git :commit => "-aqm 'Add default stuff.'"