+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - ask for user credentials before creating the Heroku app
 - set the correct timezone config on heroku after creating the app
 - move RVM config from .rvmrc to .ruby-version and .ruby-gemset files
+- refactor `integration.rake` and `deploy.rake` to handle projects with staging and production apps
 
 ### bug fixes